Initial Commit
This is a working - mostly - script that looks at a defined /24 subnet
and will then determine how many leases are in the scope as well as the
number of used leases and available leases.  The problem is that the
DHCP results include expired leases and therefore the script
over-reports used leases.